new private method _longRound; _stdev method now does all math in long, and returns long instead of float; mean and median methods refactored to use _longRound --KW

+++ tests/HundredNumberTest/HundredNumberTest.ino
@@ -30,7 +30,7 @@ void loop() {
   int median = filteredData.median(); 
   int minimum = filteredData.minimum(); 
   int maximum = filteredData.maximum(); 
-  float stdev = filteredData.stdev(); 
+  int stdev = filteredData.stdev(); 
 
   // output the results to the serial port so you can see them
   Serial.print("  mean is ");
